hello\

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: rtamouss <rtamouss@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/19 11:57:31 by rtamouss          #+#    #+#             */
-/*   Updated: 2023/11/20 22:16:02 by rtamouss         ###   ########.fr       */
+/*   Updated: 2023/11/21 10:28:19 by rtamouss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,6 @@
 # include <stdarg.h>
 # include "../libft/libft.h"
 
-void	ft_putnbr_x(unsigned int nbr, char c);
+void	ft_putnbr_x(unsigned long long nbr, char c);
 
 #endif
